fix: move utils to core/utils
