+++ src/viz/dataview/mode/DataViewLocal.ts
@@ -4,7 +4,7 @@ import { AggregationType, aggregateValues } from '@/data/operations/aggregation'
 import { groupValuesByColumn } from '@/data/operations/grouping';
 import { castToNumberOrUndefined } from '@/core/utils/number';
 import { ColumnFilters, SpatialFilters } from '@/viz/filters/types';
-import { GenericDataSourceEvent } from '@/viz/utils';
+import { GenericDataSourceEvent } from '@/viz/core/utils';
 import { DataViewMode } from './DataViewMode';
 import { GetDataOptions } from '../DataViewImpl';
 import { DataViewEvent } from '../utils';
